+++ Makefile
@@ -6,7 +6,7 @@ GMOCK_DIR=googletest/googlemock
 GTEST_FLAGS=-std=c++11 -isystem $(GTEST_DIR)/include 
 GMOCK_FLAGS=-isystem $(GMOCK_DIR)/include
 CXXFLAGS= -g $(CXXOPTIMIZE) -Wall -Werror -pedantic -std=c++11 $(BOOST)
-CLASSES=nginx-configparser/config_parser server/server webserver server/httpRequest
+CLASSES=nginx-configparser/config_parser server/server webserver server/httpRequest webserver_options
 GCOV=config_parser.cc server.cc webserver.cc
 UTIL_CLASSES=$(CLASSES:=.cc)
 TESTS=$(CLASSES:=_test.cc)
@@ -20,6 +20,7 @@ nginx-configparser/config_parser.cc: nginx-configparser/config_parser.h
 server/server.cc: server/server.h
 webserver.cc: webserver.h
 server/httpRequest.cc: server/httpRequest.h
+webserver_options.cc: webserver_options.h
 
 webserver: $(UTIL_CLASSES)
 	$(CXX) -o $@ $^ $(CXXFLAGS) main.cc
