+++ webserver.cc
@@ -1,39 +1,11 @@
 #include "webserver.h"
+#include "webserver_options.h"
 #include <utility>
 #include <map>
 #include <numeric>
 #include <boost/asio.hpp>
 
 
-std::string WebserverOptions::ToString() const{
-  std::string options_string;
-  for(auto& option : *options_) {
-    for(auto& vals : option) {
-      options_string.append(vals.first);
-      options_string.append(": ");
-      for(auto& token : vals.second) {
-        options_string.append(token);
-        options_string.append(" ");
-      }
-      options_string.append("\n");
-    }
-    
-  }
-  // printf("%s", options_string.c_str());
-  return options_string;
-}
-
-WebserverOptions::WebserverOptions(std::unique_ptr<NginxConfig> const &statement, std::vector<std::map<std::string, std::vector<std::string> > >* options) : options_(options) {
-    if(statement.get() != nullptr) {
-      for(const auto& s : statement->statements_) {
-        std::string key = s->tokens_[0];
-        std::vector<std::string> vals = std::vector<std::string>(s->tokens_.begin() + 1, s->tokens_.end());
-        std::map<std::string, std::vector<std::string> > map;
-        map.insert(std::make_pair(key, vals));
-        options_->push_back(map);
-      }
-    }
-}
 
 
 std::string Webserver::ToString() const{
@@ -110,7 +82,7 @@ bool Webserver::run_server(const char* file_name) {
   }
   try {  
     boost::asio::io_service io_service;
-    Server s(io_service, port_);
+    Server s(io_service, port_, &options_);
     printf("Running server on port %d...\n", port_);
     io_service.run();
   } 
@@ -120,6 +92,3 @@ bool Webserver::run_server(const char* file_name) {
   }
   return true;
 }
-
-
-
